Use explicit cast to avoid clang complains on implict one.
clamfs.cxx:1149:21: error: implicit conversion loses integer precision:
'uint64_t' (aka 'unsigned long') to 'int' [-Werror,-Wshorten-64-to-32]
    res = lseek(fi->fh, off, whence);
          ~~~~~ ~~~~^~

+++ src/clamfs.cxx
@@ -1146,7 +1146,7 @@ static off_t clamfs_lseek(const char *path, off_t off, int whence,
     off_t res;
     (void) path;
 
-    res = lseek(fi->fh, off, whence);
+    res = lseek((int)fi->fh, off, whence);
     if (res == -1)
         return -errno;
